+++ flexmeasures/api/v3_0/__init__.py
@@ -44,6 +44,7 @@ from flexmeasures.data.schemas.automations import (
     AutomationUpdateSchema,
 )
 from flexmeasures.data.schemas.generic_assets import GenericAssetSchema as AssetSchema
+from flexmeasures.data.schemas.reporting import ReportTriggerSchema
 from flexmeasures.data.schemas.sensors import QuantitySchema, TimeSeriesSchema
 from flexmeasures.data.schemas.account import AccountSchema
 from flexmeasures.api.v3_0.accounts import AccountAPIQuerySchema
@@ -163,6 +164,7 @@ def create_openapi_specs(app: Flask):
         ("AnnotationSchema", AnnotationSchema),
         ("AutomationCreationSchema", AutomationCreationSchema),
         ("AutomationUpdateSchema", AutomationUpdateSchema),
+        ("ReportTriggerSchema", ReportTriggerSchema),
         ("CopyAssetSchema", CopyAssetSchema),
         ("DefaultAssetViewJSONSchema", DefaultAssetViewJSONSchema),
         ("AccountSchema", AccountSchema(partial=True)),